check if mongodb is installed, before start mongodb-hack

+++ travis.sh
@@ -30,8 +30,9 @@ sudo apt-get install -qq -y python-catkin-pkg python-rosdep python-wstool ros-$R
 if [ $ROSWS == rosws ]; then sudo apt-get install -qq -y python-rosinstall     ; fi
 if [ $BUILDER == rosbuild ]; then sudo apt-get install -qq -y ros-$ROS_DISTRO-rosmake ; fi
 # MongoDB hack - I don't fully understand this but its for moveit_warehouse
-sudo apt-get remove -y mongodb mongodb-10gen || echo "ok"
-sudo apt-get install -y mongodb-clients mongodb-server -o Dpkg::Options::="--force-confdef" || ecoh "ok" # default actions
+(dpkg -s mongodb || echo "ok"; export HAVE_MONGO_DB=$?)
+if [ $HAVE_MONG_DB == 0 ]; then sudo apt-get remove -y mongodb mongodb-10gen || echo "ok"; fi
+if [ $HAVE_MONG_DB == 0 ]; then sudo apt-get install -y mongodb-clients mongodb-server -o Dpkg::Options::="--force-confdef" || echo "ok"; fi # default actions
 # Setup rosdep
 sudo rosdep init
 rosdep update; while [ $? != 0 ]; do sleep 1; rosdep update; done
